% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_images_setup.R
\name{compress_images_available}
\alias{compress_images_available}
\title{Is 'compress-images' available}
\usage{
compress_images_available()
}
\value{
a single logical value.
}
\description{
Checks if 'compress-images' is available within a directory of the user.
}
\examples{
compress_images_available()
}
\seealso{
Other tools for 'compress-images': 
\code{\link{compress_images_install}()},
\code{\link{compress_images_uninstall}()}
}
\concept{tools for 'compress-images'}
