% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_images_setup.R
\name{compress_images_install}
\alias{compress_images_install}
\title{Install 'compress-images'}
\usage{
compress_images_install(force = FALSE, verbose = TRUE)
}
\arguments{
\item{force}{Whether to force to install (override) 'compress-images'.}

\item{verbose}{should a log be printed in the console, default to TRUE.}
}
\value{
a single logical value, FALSE if the operation failed, TRUE otherwise.

a single logical value, TRUE if success.
}
\description{
Downloads and installs 'compress-images'
(a "JavaScript" tool for image compression) in the user data directory.
}
\examples{
library(locatexec)
if(exec_available("node") && !compress_images_available()){
  compress_images_install()
  compress_images_uninstall()
}
}
\seealso{
Other tools for 'compress-images': 
\code{\link{compress_images_available}()},
\code{\link{compress_images_uninstall}()}
}
\concept{tools for 'compress-images'}
