% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_images_setup.R
\name{compress_images_uninstall}
\alias{compress_images_uninstall}
\title{Uninstall 'compress-images'}
\usage{
compress_images_uninstall()
}
\value{
a single logical value, FALSE if the operation failed, TRUE otherwise.

a single logical value, TRUE if success.
}
\description{
Removes 'compress-images'.
}
\examples{
library(locatexec)

if(exec_available("npm") &&
   compress_images_available()) {
  compress_images_uninstall()
  compress_images_install()
}
}
\seealso{
Other tools for 'compress-images': 
\code{\link{compress_images_available}()},
\code{\link{compress_images_install}()}

Other tools for 'compress-images': 
\code{\link{compress_images_available}()},
\code{\link{compress_images_install}()}
}
\concept{tools for 'compress-images'}
