% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_images_setup.R
\name{get_minimage_defaults}
\alias{get_minimage_defaults}
\title{Get minimage defaults settings}
\usage{
get_minimage_defaults()
}
\value{
a list containing default values.
}
\description{
The current settings are returned by this function.
}
\examples{
get_minimage_defaults()
}
