% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_images_setup.R
\name{set_minimage_defaults}
\alias{set_minimage_defaults}
\title{Modify minimage defaults settings}
\usage{
set_minimage_defaults(compimg_path = NULL)
}
\arguments{
\item{compimg_path}{compress image node project installed path.
the original project is located at \code{system.file(package = "minimage", "compress-images")}.}
}
\description{
The current settings are automatically used to
find compress image utility path for example. This function let you
set the values of these settings.
}
